fix: update quick start docs
